% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tes0020.R
\name{pdf_worksheet}
\alias{pdf_worksheet}
\title{tes0020 PDF Template}
\usage{
pdf_worksheet(
  logo = system.file("rmarkdown/templates/tes0020/resources/taltech.pdf", package =
    "tes0020"),
  institution = NULL,
  short_title = NULL,
  watermark = NULL,
  libertine = TRUE,
  logo_height = "4cm",
  watermark_color = "gray",
  footer_on_first_page = TRUE,
  toc = FALSE,
  lot = FALSE,
  lof = FALSE,
  fancy_captions = TRUE,
  number_sections = TRUE,
  latex_engine = "xelatex",
  ...
)
}
\arguments{
\item{logo}{File path to institution logo file. The logo will be printed on
the top right corner of the first page and the size will be adjusted to
`logo_height`. The default is the TalTech logo.}

\item{institution}{Information about the institution as the page footnote. You can input a list of
values.}

\item{short_title}{A shorter version of the title to be printed on the
top-right corner of every page, except the front page.}

\item{watermark}{Watermark text to be printed on the page. You can change
its color via `watermark_color`.}

\item{libertine}{T/F. Libertine is a collection of open fonts for western
languages. Default is `TRUE`.}

\item{logo_height}{Height of the logo image. This logo image will be scaled
to height and the default height is 4cm.}

\item{watermark_color}{Color for the watermark text. Default is "gray".}

\item{footer_on_first_page}{T/F value for whether the institution info footer 
will be displayed on the front page.}

\item{toc}{T/F value for table of contents. See ?rmarkdown::pdf_document
for details}

\item{lot}{T/F value for list of tables.}

\item{lof}{T/F value for list of figures.}

\item{fancy_captions}{T/F value for whether to apply special formatting to
captions.}

\item{number_sections}{T/F value for whether sections should be numbered.
See ?rmarkdown::pdf_document for details}

\item{latex_engine}{LaTeX engine. See ?rmarkdown::pdf_document for details}

\item{...}{other options to be passed to rmarkdown::pdf_document. See
?rmarkdown::pdf_document for details}
}
\description{
tes0020 PDF Template
}
